Improve the reliability of some tests on Linux
This marks a couple tests that run `ps` as requiring that executable, since it's not normally installed by default in a Docker container. It also accounts for the fact that `ulimit -n` can be 2^20 in a Docker container, so we constrain concurrently spawned processes to an upper bound of 4096 instead.