Initial CMake Build
Adding an initial CMake build for the Swift-subprocess library.
At this time, it will build its own swift-system library, although we
should allow it to search for one if one is available.

CMake does not automatically generate modulemaps like SwiftPM does. As a
result, we need to add one for Swift to import the C shims.

There is one library target emitted from this build, Subprocess.