Improve reliability of testSuspendResumeProcess
It's valid for a process to end up in a sleep state between transitions, so don't fail the test in that case. Only fail if the process prematurely transitions to a zombie state, which means it unexpectedly exited. This fixes an issue where testSuspendResumeProcess sometimes nondeterministically fails when the process transitions to a sleep state after resuming from stopped.