Fix remaining Windows test issues
testPlatformOptionsCreateNewConsole: this test was meaning to assert that the values are different, not that they're the same. Fix this simple logic issue.
testRunDetached: this test never completed reading the output, because the write fd was kept open and reading never completed. Close the write fd before reading.
testInputSequenceCustomExecutionBody, testInputAsyncSequenceCustomExecutionBody, testRedirectedOutputRedirectToSequence: these tests hit an implementation issue where EOF was not correctly detected, leading to a hang. Change the logic to return nil if an empty buffer is received.

This patch also enables CI tests for Windows.