Introudce CombinedErrorOutput (#180)
* Introudce CombinedErrorOutput

CombinedErrorOutput is a concrete error output type for subprocesses that combines the standard error
output with the standard output stream.

When `CombinedErrorOutput` is used as the error output for a subprocess, both
standard output and standard error from the child process are merged into a
single output stream. This is equivalent to using shell redirection like `2>&1`.

* Remove Ubuntu Focal from testing targets for Linux

Swift 6.2 dropped support for Ubuntu 20.04 Focal